Software/Python_app/requirements.txt
pyqt6
opencv-python
pyserial
mouse
numpy
